rag/pyproject.toml
[project]
name = "app"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
  "torch==2.4.1",
  "torchvision==0.19.1",
  "torchaudio==2.4.1",
  "chromadb>=1.1.0",
  "sentence-transformers>=2.7.0",
  "google-cloud-storage>=2.18.0",
  "tiktoken>=0.8.0",  # optional but used by preprocess for token counts
]

# Tell uv to install torch-family packages from the PyTorch CPU index only
[tool.uv.sources]
torch = { index = "pytorch" }
torchvision = { index = "pytorch" }
torchaudio = { index = "pytorch" }

# Define the CPU index (note the double brackets — it's an array of tables)
[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cpu"
explicit = true
